added view payout instructions added faucet information added view pool login link added points success (percent) fixed bugs

+++ src/components/plotNFT/PlotNFTCard.tsx
@@ -14,6 +14,7 @@ import {
   Loading,
   FormatLargeNumber,
   Link,
+  ConfirmDialog,
 } from '@chia/core';
 import {
   Box,
@@ -25,17 +26,25 @@ import {
   MenuItem,
   ListItemIcon,
 } from '@material-ui/core';
-import { Delete as DeleteIcon } from '@material-ui/icons';
+import {
+  Delete as DeleteIcon,
+  Link as LinkIcon,
+  Payment as PaymentIcon,
+} from '@material-ui/icons';
 import type PlotNFT from '../../types/PlotNFT';
 import PlotNFTName from './PlotNFTName';
 import PlotNFTStatus from './PlotNFTState';
 import PlotIcon from '../icons/Plot';
 import usePlotNFTDetails from '../../hooks/usePlotNFTDetails';
+import useOpenDialog from '../../hooks/useOpenDialog';
 import PoolJoin from '../pool/PoolJoin';
 import PoolAbsorbRewards from '../pool/PoolAbsorbRewards';
 import { mojo_to_chia } from '../../util/chia';
 import { deleteUnconfirmedTransactions } from '../../modules/incoming';
 import PlotNFTGraph from './PlotNFTGraph';
+import PlotNFTGetPoolLoginLinkDialog from './PlotNFTGetPoolLoginLinkDialog';
+import PlotNFTPayoutInstructionsDialog from './PlotNFTPayoutInstructionsDialog';
+import getPercentPointsSuccessfull from '../../util/getPercentPointsSuccessfull';
 
 const StyledCard = styled(Card)`
   display: flex;
@@ -72,14 +81,21 @@ export default function PlotNFTCard(props: Props) {
     nft: {
       pool_state: {
         p2_singleton_puzzle_hash,
-        pool_config: { launcher_id, pool_url },
+        pool_config: {
+          launcher_id,
+          pool_url,
+        },
         points_found_24h,
+        points_acknowledged_24h,
       },
       pool_wallet_status: { wallet_id },
     },
   } = props;
 
+  const percentPointsSuccessful24 = getPercentPointsSuccessfull(points_acknowledged_24h, points_found_24h);
+
   const history = useHistory();
+  const openDialog = useOpenDialog();
   const dispatch = useDispatch();
   const { isSelfPooling, isSynced, plots, balance } = usePlotNFTDetails(nft);
   const totalPointsFound24 = points_found_24h.reduce(
@@ -96,8 +112,30 @@ export default function PlotNFTCard(props: Props) {
     });
   }
 
-  function handleDeleteUnconfirmedTransactions() {
-    dispatch(deleteUnconfirmedTransactions(wallet_id));
+  async function handleDeleteUnconfirmedTransactions() {
+    const deleteConfirmed = await openDialog(
+      <ConfirmDialog
+        title={<Trans>Confirmation</Trans>}
+        confirmTitle={<Trans>Delete</Trans>}
+        confirmColor="danger"
+      >
+        <Trans>
+          Are you sure you want to delete unconfirmed transactions?
+        </Trans>
+      </ConfirmDialog>,
+    );
+
+    if (deleteConfirmed) {
+      dispatch(deleteUnconfirmedTransactions(wallet_id));
+    }
+  }
+
+  function handleGetPoolLoginLink() {
+    openDialog(<PlotNFTGetPoolLoginLinkDialog nft={nft} />);
+  }
+
+  function handlePayoutInstructions() {
+    openDialog(<PlotNFTPayoutInstructionsDialog nft={nft} />);
   }
 
   const rows = [
@@ -190,6 +228,20 @@ export default function PlotNFTCard(props: Props) {
       ),
       value: <FormatLargeNumber value={totalPointsFound24} />,
     },
+    !isSelfPooling && {
+      key: 'points_found_24',
+      label: (
+        <Typography>
+          <Trans>Points Successful in Last 24 Hours</Trans>
+        </Typography>
+      ),
+      value: (
+        <>
+          <FormatLargeNumber value={Number(percentPointsSuccessful24 * 100).toFixed(2)} />
+          {' %'}
+        </>
+      ),
+    },
   ].filter((row) => !!row);
 
   return (
@@ -217,6 +269,36 @@ export default function PlotNFTCard(props: Props) {
                         <Trans>Add a Plot</Trans>
                       </Typography>
                     </MenuItem>
+                    {!isSelfPooling && (
+                      <MenuItem
+                        onClick={() => {
+                          onClose();
+                          handleGetPoolLoginLink();
+                        }}
+                      >
+                        <ListItemIcon>
+                          <LinkIcon />
+                        </ListItemIcon>
+                        <Typography variant="inherit" noWrap>
+                          <Trans>View Pool Login Link</Trans>
+                        </Typography>
+                      </MenuItem>
+                    )}
+                    {!isSelfPooling && (
+                      <MenuItem
+                        onClick={() => {
+                          onClose();
+                          handlePayoutInstructions();
+                        }}
+                      >
+                        <ListItemIcon>
+                          <PaymentIcon />
+                        </ListItemIcon>
+                        <Typography variant="inherit" noWrap>
+                          <Trans>View Payout Instructions</Trans>
+                        </Typography>
+                      </MenuItem>
+                    )}
                     <MenuItem
                       onClick={() => {
                         onClose();
