+++ src/components/wallet/standard/WalletStandard.tsx
@@ -10,6 +10,7 @@ import {
   CopyToClipboard,
   Flex,
   Card,
+  ConfirmDialog,
 } from '@chia/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { useDispatch, useSelector } from 'react-redux';
@@ -52,6 +53,7 @@ import { deleteUnconfirmedTransactions } from '../../../modules/incoming';
 // import WalletGraph from '../WalletGraph';
 import WalletCards from './WalletCards';
 import WalletStatus from '../WalletStatus';
+import useOpenDialog from '../../../hooks/useOpenDialog';
 
 const drawerWidth = 240;
 
@@ -394,6 +396,7 @@ function SendCard(props: SendCardProps) {
   const { sending_transaction, send_transaction_result } = wallet;
 
   const result = get_transaction_result(send_transaction_result);
+
   const result_message = result.message;
   const result_class = result.success
     ? classes.resultSuccess
@@ -617,9 +620,25 @@ type StandardWalletProps = {
 export default function StandardWallet(props: StandardWalletProps) {
   const { wallet_id } = props;
   const dispatch = useDispatch();
+  const openDialog = useOpenDialog();
 
-  function handleDeleteUnconfirmedTransactions() {
-    dispatch(deleteUnconfirmedTransactions(wallet_id));
+  async function handleDeleteUnconfirmedTransactions() {
+    const deleteConfirmed = await openDialog(
+      <ConfirmDialog
+        title={<Trans>Confirmation</Trans>}
+        confirmTitle={<Trans>Delete</Trans>}
+        confirmColor="danger"
+      >
+        <Trans>
+          Are you sure you want to delete unconfirmed transactions?
+        </Trans>
+      </ConfirmDialog>,
+    );
+
+    // @ts-ignore
+    if (deleteConfirmed) {
+      dispatch(deleteUnconfirmedTransactions(wallet_id));
+    }
   }
 
   return (
